@@ -1,27 +1,16 @@
-// +build !pcap
-// +build !rawsocket
-// +build !windivert
+//go:build !pcap && !rawsocket && !windivert
+// +build !pcap,!rawsocket,!windivert
 
 package phantomtcp
 
 func DevicePrint() {
 }
 
-func connectionMonitor(device string) {
-}
-
-func udpMonitor(device string) {
-}
-
 func ConnectionMonitor(devices []string) bool {
 	return false
 }
 
-func UDPMonitor(devices []string) bool {
-	return false
-}
-
-func ModifyAndSendPacket(connInfo *ConnectionInfo, payload []byte, method uint32, ttl uint8, count int) error {
+func ModifyAndSendPacket(connInfo *ConnectionInfo, payload []byte, hint uint32, ttl uint8, count int) error {
 	return nil
 }
 
