@@ -13,6 +13,8 @@ import (
 	"github.com/google/gopacket/pcap"
 )
 
+var ConnWait4 [65536]uint32
+var ConnWait6 [65536]uint32
 var pcapHandle *pcap.Handle
 
 func DevicePrint() {
@@ -71,10 +73,10 @@ func connectionMonitor(device string) {
 		case *layers.IPv4:
 			var srcPort layers.TCPPort
 			var synAddr string
-			var method uint32 = 0
+			var hint uint32 = 0
 			if synack {
-				method = ConnWait4[tcp.DstPort]
-				if method == 0 {
+				hint = ConnWait4[tcp.DstPort]
+				if hint == 0 {
 					continue
 				}
 				srcPort = tcp.DstPort
@@ -87,11 +89,11 @@ func connectionMonitor(device string) {
 				result, ok := ConnSyn.Load(synAddr)
 				if ok {
 					info := result.(SynInfo)
-					method = info.Option
+					hint = info.Option
 				}
 			}
 
-			if method != 0 {
+			if hint != 0 {
 				if synack {
 					srcIP := ip.DstIP
 					ip.DstIP = ip.SrcIP
@@ -119,9 +121,9 @@ func connectionMonitor(device string) {
 					connInfo = &ConnectionInfo{nil, ip, *tcp}
 				}
 
-				if method&(OPT_TFO|OPT_HTFO|OPT_SYNX2) != 0 {
+				if hint&(OPT_TFO|OPT_HTFO|OPT_SYNX2) != 0 {
 					if synack {
-						if method&(OPT_TFO|OPT_HTFO) != 0 {
+						if hint&(OPT_TFO|OPT_HTFO) != 0 {
 							for _, op := range tcp.Options {
 								if op.OptionType == 34 {
 									TFOCookies.Store(ip.DstIP.String(), op.OptionData)
@@ -129,10 +131,10 @@ func connectionMonitor(device string) {
 							}
 						}
 						ConnWait4[srcPort] = 0
-					} else if method&(OPT_TFO|OPT_HTFO) != 0 {
+					} else if hint&(OPT_TFO|OPT_HTFO) != 0 {
 						if ip.TTL < 64 {
 							count := 1
-							if method&OPT_SYNX2 != 0 {
+							if hint&OPT_SYNX2 != 0 {
 								count = 2
 							}
 
@@ -142,7 +144,7 @@ func connectionMonitor(device string) {
 								if payload != nil {
 									ip.TOS = 0
 									ModifyAndSendPacket(connInfo, payload, OPT_TFO, 0, count)
-									ConnWait4[srcPort] = method
+									ConnWait4[srcPort] = hint
 								} else {
 									connInfo = nil
 								}
@@ -152,7 +154,7 @@ func connectionMonitor(device string) {
 								connInfo = nil
 							}
 						}
-					} else if method&OPT_SYNX2 != 0 {
+					} else if hint&OPT_SYNX2 != 0 {
 						SendPacket(packet)
 					}
 				}
@@ -167,10 +169,10 @@ func connectionMonitor(device string) {
 		case *layers.IPv6:
 			var srcPort layers.TCPPort
 			var synAddr string
-			var method uint32 = 0
+			var hint uint32 = 0
 			if synack {
-				method = ConnWait6[tcp.DstPort]
-				if method == 0 {
+				hint = ConnWait6[tcp.DstPort]
+				if hint == 0 {
 					continue
 				}
 				srcPort = tcp.DstPort
@@ -183,10 +185,10 @@ func connectionMonitor(device string) {
 				result, ok := ConnSyn.Load(synAddr)
 				if ok {
 					info := result.(SynInfo)
-					method = info.Option
+					hint = info.Option
 				}
 			}
-			if method != 0 {
+			if hint != 0 {
 				if synack {
 					srcIP := ip.DstIP
 					ip.DstIP = ip.SrcIP
@@ -214,9 +216,9 @@ func connectionMonitor(device string) {
 					connInfo = &ConnectionInfo{nil, ip, *tcp}
 				}
 
-				if method&(OPT_TFO|OPT_HTFO|OPT_SYNX2) != 0 {
+				if hint&(OPT_TFO|OPT_HTFO|OPT_SYNX2) != 0 {
 					if synack {
-						if method&(OPT_TFO|OPT_HTFO) != 0 {
+						if hint&(OPT_TFO|OPT_HTFO) != 0 {
 							for _, op := range tcp.Options {
 								if op.OptionType == 34 {
 									TFOCookies.Store(ip.DstIP.String(), op.OptionData)
@@ -224,10 +226,10 @@ func connectionMonitor(device string) {
 							}
 						}
 						ConnWait6[srcPort] = 0
-					} else if method&(OPT_TFO|OPT_HTFO) != 0 {
+					} else if hint&(OPT_TFO|OPT_HTFO) != 0 {
 						if ip.HopLimit < 64 {
 							count := 1
-							if method&OPT_SYNX2 != 0 {
+							if hint&OPT_SYNX2 != 0 {
 								count = 2
 							}
 
@@ -237,7 +239,7 @@ func connectionMonitor(device string) {
 								if payload != nil {
 									ip.TrafficClass = 0
 									ModifyAndSendPacket(connInfo, payload, OPT_TFO, 0, count)
-									ConnWait4[srcPort] = method
+									ConnWait4[srcPort] = hint
 								} else {
 									connInfo = nil
 								}
@@ -247,7 +249,7 @@ func connectionMonitor(device string) {
 								connInfo = nil
 							}
 						}
-					} else if method&OPT_SYNX2 != 0 {
+					} else if hint&OPT_SYNX2 != 0 {
 						SendPacket(packet)
 					}
 				}
@@ -258,51 +260,6 @@ func connectionMonitor(device string) {
 					case <-time.After(time.Second * 2):
 					}
 				}(connInfo)
-			}
-		}
-	}
-}
-
-func udpMonitor(device string) {
-	fmt.Printf("Device: %v (UDP)\n", device)
-
-	snapLen := int32(65535)
-
-	filter := "net 6.0.0.0/8 and udp port 443"
-
-	var err error
-	pcapHandle, err = pcap.OpenLive(device, snapLen, true, pcap.BlockForever)
-	if err != nil {
-		fmt.Printf("pcap open live failed: %v", err)
-		return
-	}
-
-	if err = pcapHandle.SetBPFFilter(filter); err != nil {
-		fmt.Printf("set bpf filter failed: %v", err)
-		return
-	}
-	defer pcapHandle.Close()
-
-	packetSource := gopacket.NewPacketSource(pcapHandle, pcapHandle.LinkType())
-	packetSource.NoCopy = false
-	for {
-		packet, err := packetSource.NextPacket()
-		if err != nil {
-			logPrintln(1, err)
-			continue
-		}
-
-		//link := packet.LinkLayer()
-		network := packet.NetworkLayer()
-		transport := packet.TransportLayer()
-
-		switch network := network.(type) {
-		case *layers.IPv4:
-			switch transport := transport.(type) {
-			case *layers.UDP:
-				src := net.UDPAddr{IP: network.SrcIP, Port: int(transport.SrcPort)}
-				dst := net.UDPAddr{IP: network.DstIP, Port: int(transport.DstPort)}
-				logPrintln(1, src, "->", dst)
 			}
 		}
 	}
@@ -321,14 +278,6 @@ func ConnectionMonitor(devices []string) bool {
 
 	for i := 0; i < len(devices); i++ {
 		go connectionMonitor(devices[i])
-	}
-
-	return true
-}
-
-func UDPMonitor(devices []string) bool {
-	for i := 0; i < len(devices); i++ {
-		go udpMonitor(devices[i])
 	}
 
 	return true
