@@ -6,7 +6,7 @@ import (
 	"time"
 )
 
-func DialConnInfo(laddr, raddr *net.TCPAddr, server *PhantomServer, payload []byte) (net.Conn, *ConnectionInfo, error) {
+func DialConnInfo(laddr, raddr *net.TCPAddr, server *PhantomInterface, payload []byte) (net.Conn, *ConnectionInfo, error) {
 	var conn net.Conn
 	var err error
 
@@ -23,19 +23,19 @@ func DialConnInfo(laddr, raddr *net.TCPAddr, server *PhantomServer, payload []by
 		}()
 	}
 
-	AddConn(addr, server.Option)
+	AddConn(addr, server.Hint)
 
-	if (server.Option & (OPT_MSS | OPT_TFO | OPT_HTFO | OPT_KEEPALIVE)) != 0 {
+	if (server.Hint & (OPT_MSS | OPT_TFO | OPT_HTFO | OPT_KEEPALIVE)) != 0 {
 		d := net.Dialer{Timeout: timeout, LocalAddr: laddr,
 			Control: func(network, address string, c syscall.RawConn) error {
 				err := c.Control(func(fd uintptr) {
 					f := syscall.Handle(fd)
-					if (server.Option & OPT_MSS) != 0 {
+					if (server.Hint & OPT_MSS) != 0 {
 					}
-					if (server.Option & (OPT_TFO | OPT_HTFO)) != 0 {
+					if (server.Hint & (OPT_TFO | OPT_HTFO)) != 0 {
 						syscall.SetsockoptInt(f, syscall.IPPROTO_IP, syscall.IP_TTL, tfo_id|64)
 					}
-					if (server.Option & OPT_KEEPALIVE) != 0 {
+					if (server.Hint & OPT_KEEPALIVE) != 0 {
 						syscall.SetsockoptInt(f, syscall.SOL_SOCKET, syscall.SO_KEEPALIVE, 1)
 					}
 				})
@@ -110,7 +110,7 @@ func GetOriginalDST(conn *net.TCPConn) (*net.TCPAddr, error) {
 
 	if ip4 := LocalTCPAddr.IP.To4(); ip4 != nil {
 		if ip4[0] == 127 && ip4[1] == 255 {
-			ip4[0] = 6
+			ip4[0] = VirtualAddrPrefix
 			ip4[1] = 0
 			LocalTCPAddr.IP = ip4
 			RemoteTCPAddr := conn.RemoteAddr().(*net.TCPAddr).IP.To4()
@@ -119,4 +119,8 @@ func GetOriginalDST(conn *net.TCPConn) (*net.TCPAddr, error) {
 	}
 
 	return LocalTCPAddr, nil
+}
+
+func SendWithOption(conn net.Conn, payload []byte, tos, ttl int) error {
+	return nil
 }
