@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"net"
 	"strconv"
-	"strings"
 	"sync"
 	"time"
 
@@ -17,6 +16,8 @@ import (
 	"github.com/macronut/godivert"
 )
 
+var ConnWait4 [65536]uint32
+var ConnWait6 [65536]uint32
 var winDivertLock sync.Mutex
 var winDivert *godivert.WinDivertHandle
 
@@ -59,10 +60,10 @@ func connectionMonitor(layer uint8) {
 		case *layers.IPv4:
 			var srcPort layers.TCPPort
 			var synAddr string
-			var method uint32 = 0
+			var hint uint32 = 0
 			if synack {
-				method = ConnWait4[tcp.DstPort]
-				if method == 0 {
+				hint = ConnWait4[tcp.DstPort]
+				if hint == 0 {
 					winDivert.Send(divertpacket)
 					continue
 				}
@@ -76,11 +77,11 @@ func connectionMonitor(layer uint8) {
 				result, ok := ConnSyn.Load(synAddr)
 				if ok {
 					info := result.(SynInfo)
-					method = info.Option
+					hint = info.Option
 				}
 			}
 
-			if method != 0 {
+			if hint != 0 {
 				if synack {
 					srcIP := ip.DstIP
 					ip.DstIP = ip.SrcIP
@@ -97,9 +98,9 @@ func connectionMonitor(layer uint8) {
 				ch := ConnInfo4[srcPort]
 				connInfo := &ConnectionInfo{nil, ip, *tcp}
 
-				if method&(OPT_TFO|OPT_HTFO|OPT_SYNX2) != 0 {
+				if hint&(OPT_TFO|OPT_HTFO|OPT_SYNX2) != 0 {
 					if synack {
-						if method&(OPT_TFO|OPT_HTFO) != 0 {
+						if hint&(OPT_TFO|OPT_HTFO) != 0 {
 							for _, op := range tcp.Options {
 								if op.OptionType == 34 {
 									TFOCookies.Store(ip.DstIP.String(), op.OptionData)
@@ -107,10 +108,10 @@ func connectionMonitor(layer uint8) {
 							}
 						}
 						ConnWait4[srcPort] = 0
-					} else if method&(OPT_TFO|OPT_HTFO) != 0 {
+					} else if hint&(OPT_TFO|OPT_HTFO) != 0 {
 						if ip.TTL < 128 {
 							count := 1
-							if method&OPT_SYNX2 != 0 {
+							if hint&OPT_SYNX2 != 0 {
 								count = 2
 							}
 
@@ -121,7 +122,7 @@ func connectionMonitor(layer uint8) {
 								if payload != nil {
 									ip.TOS = 0
 									ModifyAndSendPacket(connInfo, payload, OPT_TFO, 0, count)
-									ConnWait4[srcPort] = method
+									ConnWait4[srcPort] = hint
 								} else {
 									connInfo = nil
 								}
@@ -131,7 +132,7 @@ func connectionMonitor(layer uint8) {
 								connInfo = nil
 							}
 						}
-					} else if method&OPT_SYNX2 != 0 {
+					} else if hint&OPT_SYNX2 != 0 {
 						winDivert.Send(divertpacket)
 						SendPacket(packet)
 					}
@@ -151,10 +152,10 @@ func connectionMonitor(layer uint8) {
 		case *layers.IPv6:
 			var srcPort layers.TCPPort
 			var synAddr string
-			var method uint32 = 0
+			var hint uint32 = 0
 			if synack {
-				method = ConnWait6[tcp.DstPort]
-				if method == 0 {
+				hint = ConnWait6[tcp.DstPort]
+				if hint == 0 {
 					winDivert.Send(divertpacket)
 					continue
 				}
@@ -168,10 +169,10 @@ func connectionMonitor(layer uint8) {
 				result, ok := ConnSyn.Load(synAddr)
 				if ok {
 					info := result.(SynInfo)
-					method = info.Option
+					hint = info.Option
 				}
 			}
-			if method != 0 {
+			if hint != 0 {
 				if synack {
 					srcIP := ip.DstIP
 					ip.DstIP = ip.SrcIP
@@ -188,9 +189,9 @@ func connectionMonitor(layer uint8) {
 				ch := ConnInfo6[srcPort]
 				connInfo := &ConnectionInfo{nil, ip, *tcp}
 
-				if method&(OPT_TFO|OPT_HTFO|OPT_SYNX2) != 0 {
+				if hint&(OPT_TFO|OPT_HTFO|OPT_SYNX2) != 0 {
 					if synack {
-						if method&(OPT_TFO|OPT_HTFO) != 0 {
+						if hint&(OPT_TFO|OPT_HTFO) != 0 {
 							for _, op := range tcp.Options {
 								if op.OptionType == 34 {
 									TFOCookies.Store(ip.DstIP.String(), op.OptionData)
@@ -198,10 +199,10 @@ func connectionMonitor(layer uint8) {
 							}
 						}
 						ConnWait6[srcPort] = 0
-					} else if method&(OPT_TFO|OPT_HTFO) != 0 {
+					} else if hint&(OPT_TFO|OPT_HTFO) != 0 {
 						if ip.HopLimit < 128 {
 							count := 1
-							if method&OPT_SYNX2 != 0 {
+							if hint&OPT_SYNX2 != 0 {
 								count = 2
 							}
 
@@ -212,7 +213,7 @@ func connectionMonitor(layer uint8) {
 								if payload != nil {
 									ip.TrafficClass = 0
 									ModifyAndSendPacket(connInfo, payload, OPT_TFO, 0, count)
-									ConnWait4[srcPort] = method
+									ConnWait4[srcPort] = hint
 								} else {
 									connInfo = nil
 								}
@@ -222,7 +223,7 @@ func connectionMonitor(layer uint8) {
 								connInfo = nil
 							}
 						}
-					} else if method&OPT_SYNX2 != 0 {
+					} else if hint&OPT_SYNX2 != 0 {
 						winDivert.Send(divertpacket)
 						SendPacket(packet)
 					}
@@ -256,10 +257,6 @@ func ConnectionMonitor(devices []string) bool {
 	return true
 }
 
-func UDPMonitor(devices []string) bool {
-	return true
-}
-
 func SendPacket(packet gopacket.Packet) error {
 	payload := packet.LinkLayer().LayerPayload()
 
@@ -274,11 +271,11 @@ func SendPacket(packet gopacket.Packet) error {
 	return err
 }
 
-func ModifyAndSendPacket(connInfo *ConnectionInfo, payload []byte, method uint32, ttl uint8, count int) error {
+func ModifyAndSendPacket(connInfo *ConnectionInfo, payload []byte, hint uint32, ttl uint8, count int) error {
 	ipLayer := connInfo.IP
 
 	var tcpLayer *layers.TCP
-	if method&OPT_TFO != 0 {
+	if hint&OPT_TFO != 0 {
 		tcpLayer = &connInfo.TCP
 
 		tcpLayer.Seq -= uint32(len(payload))
@@ -315,21 +312,21 @@ func ModifyAndSendPacket(connInfo *ConnectionInfo, payload []byte, method uint32
 			Window:     connInfo.TCP.Window,
 		}
 
-		if method&OPT_WMD5 != 0 {
+		if hint&OPT_WMD5 != 0 {
 			tcpLayer.Options = []layers.TCPOption{
 				layers.TCPOption{19, 16, []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}},
 			}
-		} else if method&OPT_WTIME != 0 {
+		} else if hint&OPT_WTIME != 0 {
 			tcpLayer.Options = []layers.TCPOption{
 				layers.TCPOption{8, 8, []byte{0, 0, 0, 0, 0, 0, 0, 0}},
 			}
 		}
 	}
 
-	if method&OPT_NACK != 0 {
+	if hint&OPT_NACK != 0 {
 		tcpLayer.ACK = false
 		tcpLayer.Ack = 0
-	} else if method&OPT_WACK != 0 {
+	} else if hint&OPT_WACK != 0 {
 		tcpLayer.Ack += uint32(tcpLayer.Window)
 	}
 
@@ -338,11 +335,11 @@ func ModifyAndSendPacket(connInfo *ConnectionInfo, payload []byte, method uint32
 	var options gopacket.SerializeOptions
 	options.FixLengths = true
 
-	if method&OPT_WCSUM == 0 {
+	if hint&OPT_WCSUM == 0 {
 		options.ComputeChecksums = true
 	}
 
-	if method&OPT_WSEQ != 0 {
+	if hint&OPT_WSEQ != 0 {
 		tcpLayer.Seq--
 		fakepayload := make([]byte, len(payload)+1)
 		fakepayload[0] = 0xFF
@@ -354,14 +351,14 @@ func ModifyAndSendPacket(connInfo *ConnectionInfo, payload []byte, method uint32
 
 	switch ip := ipLayer.(type) {
 	case *layers.IPv4:
-		if method&OPT_TTL != 0 {
+		if hint&OPT_TTL != 0 {
 			ip.TTL = ttl
 		}
 		gopacket.SerializeLayers(buffer, options,
 			ip, tcpLayer, gopacket.Payload(payload),
 		)
 	case *layers.IPv6:
-		if method&OPT_TTL != 0 {
+		if hint&OPT_TTL != 0 {
 			ip.HopLimit = ttl
 		}
 		gopacket.SerializeLayers(buffer, options,
@@ -387,23 +384,24 @@ func ModifyAndSendPacket(connInfo *ConnectionInfo, payload []byte, method uint32
 	return nil
 }
 
-func SendJumboUDPPacket(laddr, raddr *net.UDPAddr, payload []byte) error {
-	return nil
-}
-
 func Redirect(dst string, to_port int, forward bool) {
 	if dst == "" {
 		return
 	}
 
 	var dstfilter string
-	iprange := strings.SplitN(dst, "-", 2)
-	logPrintln(1, dst)
-	if len(iprange) > 1 {
-		dstfilter = fmt.Sprintf("ip.DstAddr>=%s and ip.DstAddr<=%s and tcp", iprange[0], iprange[1])
+	dstip := net.ParseIP(dst).To4()
+	if dstip == nil {
+		return
+	}
+
+	if dstip[2] == 0 && dstip[3] == 0 {
+		dstfilter = fmt.Sprintf("ip.DstAddr>=%d.%d.0.0 and ip.DstAddr<%d.%d.255.255 and tcp", dstip[0], dstip[1], dstip[0], dstip[1])
 	} else {
 		dstfilter = fmt.Sprintf("ip.DstAddr=%s and tcp", dst)
 	}
+
+	logPrintln(1, dstfilter)
 
 	filter := fmt.Sprintf("(outbound and %s) or (ip.SrcAddr>127.255.0.0 and ip.SrcAddr<127.255.255.255 and tcp.SrcPort=%s)", dstfilter, strconv.Itoa(to_port))
 
@@ -464,7 +462,7 @@ func Redirect(dst string, to_port int, forward bool) {
 		dstPort, _ := packet.DstPort()
 
 		if srcIP[0] == 127 && dstIP[0] == 127 {
-			packet.SetSrcIP(net.IPv4(6, 0, srcIP[2], srcIP[3]))
+			packet.SetSrcIP(net.IPv4(dstip[0], dstip[1], srcIP[2], srcIP[3]))
 			packet.SetSrcPort(uint16(dstIP[2])<<8 | uint16(dstIP[3]))
 			if dstIP[1] > 1 {
 				packet.SetDstIP(net.IPv4(192, 168, 137, dstIP[1]))
@@ -510,27 +508,17 @@ func RedirectDNS() {
 		}
 		udpheadlen := 8
 		request := packet.Raw[ipheadlen+udpheadlen:]
-		qname, qtype, _ := GetQName(request)
+
+		qname, _, _ := GetQName(request)
 		if qname == "" {
 			logPrintln(2, "DNS Segmentation fault")
 			continue
 		}
 
-		conf, ok := ConfigLookup(qname)
-		if ok {
-			logPrintln(1, qname, conf)
-			var response []byte
-			if (conf.Option & OPT_MODIFY) != 0 {
-				index, _ := NSLookup(qname, conf.Option, conf.Server)
-				if qtype == 28 {
-					response = BuildResponse(request, qtype, 0, nil)
-				} else {
-					response = BuildLie(request, qtype, index)
-				}
-			} else {
-				response = NSRequest(request, true)
-			}
-
+		server := ConfigLookup(qname)
+		if server != nil {
+			logPrintln(1, qname, server)
+			_, response := NSRequest(request, true)
 			udpsize := len(response) + 8
 
 			var packetsize int
