@@ -30,7 +30,7 @@ func SetProxy(dev, address string, state bool) error {
 		switch u.Scheme {
 		case "redirect":
 			if state {
-				go ptcp.Redirect("6.0.0.1-6.0.255.254", proxyTCPAddr.Port, true)
+				go ptcp.Redirect(proxyTCPAddr.IP.String(), proxyTCPAddr.Port, true)
 				go ptcp.RedirectDNS()
 			}
 
